Create motion.py
Detects motion with help of a static frame at start, turn it into gray, then applies Gaussian filter on it. When a moving object comes into the picture, the same process is iterated for that one. Both frames are compared to detect moving objects and highlight a moving object by drawing a red rectangle around it.